# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LocumMap Chennai API (MVP)",
    "description": "FastAPI backend for LocumMap Chennai (MVP).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/shifts": {
      "get": {
        "tags": [
          "shifts"
        ],
        "summary": "List Shifts",
        "operationId": "list_shifts_shifts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShiftOut"
                  },
                  "type": "array",
                  "title": "Response List Shifts Shifts Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "shifts"
        ],
        "summary": "Create Shift",
        "operationId": "create_shift_shifts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/shifts/{shift_id}": {
      "get": {
        "tags": [
          "shifts"
        ],
        "summary": "Get Shift",
        "operationId": "get_shift_shifts__shift_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shift Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shifts/{shift_id}/candidates": {
      "get": {
        "tags": [
          "shifts"
        ],
        "summary": "Candidates",
        "operationId": "candidates_shifts__shift_id__candidates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shift Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateOut"
                  },
                  "title": "Response Candidates Shifts  Shift Id  Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/available": {
      "get": {
        "tags": [
          "shifts"
        ],
        "summary": "Available Jobs",
        "operationId": "available_jobs_jobs_available_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShiftOut"
                  },
                  "type": "array",
                  "title": "Response Available Jobs Jobs Available Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/shifts/{shift_id}/book": {
      "post": {
        "tags": [
          "shifts"
        ],
        "summary": "Book Shift",
        "operationId": "book_shift_shifts__shift_id__book_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shift Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftBookIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookResultOut": {
        "properties": {
          "shift_id": {
            "type": "string",
            "format": "uuid",
            "title": "Shift Id"
          },
          "assignment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assignment Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "shift_id",
          "assignment_id",
          "status"
        ],
        "title": "BookResultOut"
      },
      "CandidateOut": {
        "properties": {
          "doctor_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor User Id"
          },
          "doctor_profile_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Profile Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "specialty": {
            "type": "string",
            "title": "Specialty"
          },
          "reliability": {
            "type": "object",
            "title": "Reliability"
          },
          "eta_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Minutes"
          },
          "distance_m": {
            "type": "integer",
            "title": "Distance M"
          },
          "acceptance_likelihood": {
            "type": "number",
            "title": "Acceptance Likelihood"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "doctor_user_id",
          "doctor_profile_id",
          "full_name",
          "specialty",
          "reliability",
          "eta_minutes",
          "distance_m",
          "acceptance_likelihood",
          "score"
        ],
        "title": "CandidateOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShiftBookIn": {
        "properties": {
          "doctor_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor User Id",
            "description": "Clinic-driven booking: choose a doctor. Doctor-driven accept can omit this."
          }
        },
        "type": "object",
        "title": "ShiftBookIn"
      },
      "ShiftCreateIn": {
        "properties": {
          "specialty": {
            "$ref": "#/components/schemas/Specialty"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "pay_amount_inr": {
            "type": "integer",
            "maximum": 200000.0,
            "minimum": 100.0,
            "title": "Pay Amount Inr"
          },
          "address": {
            "type": "string",
            "maxLength": 300,
            "minLength": 5,
            "title": "Address"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "auto_replace": {
            "type": "boolean",
            "title": "Auto Replace",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "specialty",
          "start_time",
          "end_time",
          "pay_amount_inr",
          "address",
          "lat",
          "lng"
        ],
        "title": "ShiftCreateIn"
      },
      "ShiftOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "clinic_profile_id": {
            "type": "string",
            "format": "uuid",
            "title": "Clinic Profile Id"
          },
          "specialty": {
            "$ref": "#/components/schemas/Specialty"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "pay_amount_inr": {
            "type": "integer",
            "title": "Pay Amount Inr"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "auto_replace": {
            "type": "boolean",
            "title": "Auto Replace"
          },
          "status": {
            "$ref": "#/components/schemas/ShiftStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "assignment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignment Id"
          },
          "doctor_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor User Id"
          },
          "doctor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "clinic_profile_id",
          "specialty",
          "start_time",
          "end_time",
          "pay_amount_inr",
          "address",
          "lat",
          "lng",
          "notes",
          "auto_replace",
          "status",
          "created_at"
        ],
        "title": "ShiftOut"
      },
      "ShiftStatus": {
        "type": "string",
        "enum": [
          "draft",
          "posted",
          "booked",
          "en_route",
          "checked_in",
          "completed",
          "paid",
          "canceled",
          "no_show"
        ],
        "title": "ShiftStatus"
      },
...
      "Specialty": {
        "type": "string",
        "enum": [
          "dentist_general",
          "endodontist",
          "anesthetist"
        ],
        "title": "Specialty"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}